implemented Bobs comments on last commit
