+++ pyproject.toml
@@ -13,6 +13,7 @@ SQLAlchemy = "^1.4.6"
 boto3 = "^1.17.46"
 Flask = "^1.1.2"
 psycopg2-binary = "^2.8.6"
+python-dotenv = "^0.17.0"
 
 [tool.poetry.dev-dependencies]
 pytest = "^5.2"
